third_party/blink/web_tests/virtual/dialog-close-when-open-removed-disabled/external/wpt/html/semantics/interactive-elements/the-dialog-element/dialog-showModal-expected.txt

This is a testharness.js-based test.
[FAIL] when opening multiple dialogs, the most recently opened is rendered on top
  assert_equals: expected Element node <dialog id="d10" open=""></dialog> but got Element node <dialog id="d11" open=""></dialog>
Harness: the test ran to completion.
